Part 4: Specialized Workflows
