--- cse-project/backend/app/services/agent.py
@@ -1,106 +1,0 @@
-from langchain_anthropic import ChatAnthropic
-from langchain_community.tools.tavily_search import TavilySearchResults
-from langchain_core.prompts import ChatPromptTemplate
-from langchain_core.runnables import Runnable, RunnableConfig
-from app.services.state import State
-from app.core.config import settings  # Import settings to access TAVILY_API_KEY
-from datetime import datetime
-
-from app.services.tools import (
-    fetch_user_flight_information,
-    search_flights,
-    lookup_policy,
-    update_ticket_to_new_flight,
-    cancel_ticket,
-    search_car_rentals,
-    book_car_rental,
-    update_car_rental,
-    cancel_car_rental,
-    search_hotels,
-    book_hotel,
-    update_hotel,
-    cancel_hotel,
-    search_trip_recommendations,
-    book_excursion,
-    update_excursion,
-    cancel_excursion,
-)
-
-class Assistant:
-    def __init__(self, runnable: Runnable):
-        self.runnable = runnable
-
-    def __call__(self, state: State, config: RunnableConfig):
-        while True:
-            result = self.runnable.invoke(state)
-            # If the LLM happens to return an empty response, we will re-prompt it
-            # for an actual response.
-            if not result.tool_calls and (
-                not result.content
-                or isinstance(result.content, list)
-                and not result.content[0].get("text")
-            ):
-                messages = state["messages"] + [("user", "Respond with a real output.")]
-                state = {**state, "messages": messages}
-            else:
-                break
-        return {"messages": result}
-
-
-# Haiku is faster and cheaper, but less accurate
-# llm = ChatAnthropic(model="claude-3-haiku-20240307")
-
-# llm = ChatAnthropic(model="claude-3-sonnet-20240229", temperature=1, api_key=settings.ANTHROPIC_API_KEY)
-# You could swap LLMs, though you will likely want to update the prompts when
-# doing so!
-from langchain_openai import ChatOpenAI
-llm = ChatOpenAI(model="gpt-4-turbo-preview", api_key=settings.OPENAI_API_KEY)
-
-# Initialize TavilySearchResults with TAVILY_API_KEY
-
-
-assistant_prompt = ChatPromptTemplate.from_messages(
-    [
-        (
-            "system",
-            "You are a helpful customer support assistant for Swiss Airlines. "
-            " Use the provided tools to search for flights, company policies, and other information to assist the user's queries. "
-            " When searching, be persistent. Expand your query bounds if the first search returns no results. "
-            " If a search comes up empty, expand your search before giving up."
-            "\n\nCurrent user:\n<User>\n{user_info}\n</User>"
-            "\nCurrent time: {time}.",
-        ),
-        ("placeholder", "{messages}"),
-    ]
-).partial(time=datetime.now)
-# "Read"-only tools (such as retrievers) don't need a user confirmation to use
-part_3_safe_tools = [
-    #TavilySearchResults(max_results=1),
-    fetch_user_flight_information,
-    search_flights,
-    lookup_policy,
-    search_car_rentals,
-    search_hotels,
-    search_trip_recommendations,
-]
-
-# These tools all change the user's reservations.
-# The user has the right to control what decisions are made
-part_3_sensitive_tools = [
-    update_ticket_to_new_flight,
-    cancel_ticket,
-    book_car_rental,
-    update_car_rental,
-    cancel_car_rental,
-    book_hotel,
-    update_hotel,
-    cancel_hotel,
-    book_excursion,
-    update_excursion,
-    cancel_excursion,
-]
-sensitive_tool_names = {t.name for t in part_3_sensitive_tools}
-# Our LLM doesn't have to know which nodes it has to route to. In its 'mind', it's just invoking functions.
-part_3_assistant_runnable = assistant_prompt | llm.bind_tools(
-    part_3_safe_tools + part_3_sensitive_tools
-)
