+++ cse-project/backend/app/services/cs_agents/assistants/excursion_assistant.py
@@ -1,0 +1,34 @@
+from datetime import datetime
+from langchain_core.prompts import ChatPromptTemplate
+from app.services.cs_agents.assistants.common import  CompleteOrEscalate
+from app.services.cs_agents.tools import search_trip_recommendations, book_excursion, update_excursion, cancel_excursion
+from app.services.cs_agents.assistants.primary_assistant import llm
+
+book_excursion_prompt = ChatPromptTemplate.from_messages(
+    [
+        (
+            "system",
+            "You are a specialized assistant for handling trip recommendations. "
+            "The primary assistant delegates work to you whenever the user needs help booking a recommended trip. "
+            "Search for available trip recommendations based on the user's preferences and confirm the booking details with the customer. "
+            "If you need more information or the customer changes their mind, escalate the task back to the main assistant."
+            " When searching, be persistent. Expand your query bounds if the first search returns no results. "
+            " Remember that a booking isn't completed until after the relevant tool has successfully been used."
+            "\nCurrent time: {time}."
+            '\n\nIf the user needs help, and none of your tools are appropriate for it, then "CompleteOrEscalate" the dialog to the host assistant. Do not waste the user\'s time. Do not make up invalid tools or functions.'
+            "\n\nSome examples for which you should CompleteOrEscalate:\n"
+            " - 'nevermind i think I'll book separately'\n"
+            " - 'i need to figure out transportation while i'm there'\n"
+            " - 'Oh wait i haven't booked my flight yet i'll do that first'\n"
+            " - 'Excursion booking confirmed!'",
+        ),
+        ("placeholder", "{messages}"),
+    ]
+).partial(time=datetime.now)
+
+book_excursion_safe_tools = [search_trip_recommendations]
+book_excursion_sensitive_tools = [book_excursion, update_excursion, cancel_excursion]
+book_excursion_tools = book_excursion_safe_tools + book_excursion_sensitive_tools
+book_excursion_runnable = book_excursion_prompt | llm.bind_tools(
+    book_excursion_tools + [CompleteOrEscalate]
+)
