+++ cse-project/backend/app/services/cs_agents/assistants/flight_booking_assistant.py
@@ -1,0 +1,30 @@
+from datetime import datetime
+from langchain_core.prompts import ChatPromptTemplate
+from app.services.cs_agents.assistants.common import CompleteOrEscalate
+from app.services.cs_agents.tools import search_flights, update_ticket_to_new_flight, cancel_ticket
+from app.services.cs_agents.assistants.primary_assistant import llm 
+flight_booking_prompt = ChatPromptTemplate.from_messages(
+    [
+        (
+            "system",
+            "You are a specialized assistant for handling flight updates. "
+            " The primary assistant delegates work to you whenever the user needs help updating their bookings. "
+            "Confirm the updated flight details with the customer and inform them of any additional fees. "
+            " When searching, be persistent. Expand your query bounds if the first search returns no results. "
+            "If you need more information or the customer changes their mind, escalate the task back to the main assistant."
+            " Remember that a booking isn't completed until after the relevant tool has successfully been used."
+            "\n\nCurrent user flight information:\n<Flights>\n{user_info}\n</Flights>"
+            "\nCurrent time: {time}."
+            "\n\nIf the user needs help, and none of your tools are appropriate for it, then"
+            ' "CompleteOrEscalate" the dialog to the host assistant. Do not waste the user\'s time. Do not make up invalid tools or functions.',
+        ),
+        ("placeholder", "{messages}"),
+    ]
+).partial(time=datetime.now)
+
+update_flight_safe_tools = [search_flights]
+update_flight_sensitive_tools = [update_ticket_to_new_flight, cancel_ticket]
+update_flight_tools = update_flight_safe_tools + update_flight_sensitive_tools
+update_flight_runnable = flight_booking_prompt | llm.bind_tools(
+    update_flight_tools + [CompleteOrEscalate]
+)
