+++ cse-project/backend/app/services/cs_agents/assistants/primary_assistant.py
@@ -1,0 +1,117 @@
+from langchain_core.prompts import ChatPromptTemplate
+from pydantic import BaseModel, Field
+from app.services.cs_agents.tools import search_flights, lookup_policy
+from datetime import datetime
+from app.core.config import settings
+from langchain_openai import ChatOpenAI
+class ToFlightBookingAssistant(BaseModel):
+    """Transfers work to a specialized assistant to handle flight updates and cancellations."""
+
+    request: str = Field(
+        description="Any necessary followup questions the update flight assistant should clarify before proceeding."
+    )
+
+class ToBookCarRental(BaseModel):
+    """Transfers work to a specialized assistant to handle car rental bookings."""
+
+    location: str = Field(
+        description="The location where the user wants to rent a car."
+    )
+    start_date: str = Field(description="The start date of the car rental.")
+    end_date: str = Field(description="The end date of the car rental.")
+    request: str = Field(
+        description="Any additional information or requests from the user regarding the car rental."
+    )
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "location": "Basel",
+                "start_date": "2023-07-01",
+                "end_date": "2023-07-05",
+                "request": "I need a compact car with automatic transmission.",
+            }
+        }
+
+class ToHotelBookingAssistant(BaseModel):
+    """Transfer work to a specialized assistant to handle hotel bookings."""
+
+    location: str = Field(
+        description="The location where the user wants to book a hotel."
+    )
+    checkin_date: str = Field(description="The check-in date for the hotel.")
+    checkout_date: str = Field(description="The check-out date for the hotel.")
+    request: str = Field(
+        description="Any additional information or requests from the user regarding the hotel booking."
+    )
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "location": "Zurich",
+                "checkin_date": "2023-08-15",
+                "checkout_date": "2023-08-20",
+                "request": "I prefer a hotel near the city center with a room that has a view.",
+            }
+        }
+
+
+class ToBookExcursion(BaseModel):
+    """Transfers work to a specialized assistant to handle trip recommendation and other excursion bookings."""
+
+    location: str = Field(
+        description="The location where the user wants to book a recommended trip."
+    )
+    request: str = Field(
+        description="Any additional information or requests from the user regarding the trip recommendation."
+    )
+
+    class Config:
+        json_schema_extra = {
+            "example": {
+                "location": "Lucerne",
+                "request": "The user is interested in outdoor activities and scenic views.",
+            }
+        }
+
+
+# The top-level assistant performs general Q&A and delegates specialized tasks to other assistants.
+# The task delegation is a simple form of semantic routing / does simple intent detection
+# llm = ChatAnthropic(model="claude-3-haiku-20240307")
+# llm = ChatAnthropic(model="claude-3-sonnet-20240229", temperature=1)
+
+
+llm = ChatOpenAI(model="gpt-4-turbo-preview", api_key=settings.OPENAI_API_KEY)
+primary_assistant_prompt = ChatPromptTemplate.from_messages(
+    [
+        (
+            "system",
+            "You are a helpful customer support assistant for Swiss Airlines. "
+            "Your primary role is to search for flight information and company policies to answer customer queries. "
+            "If a customer requests to update or cancel a flight, book a car rental, book a hotel, or get trip recommendations, "
+            "delegate the task to the appropriate specialized assistant by invoking the corresponding tool. You are not able to make these types of changes yourself."
+            " Only the specialized assistants are given permission to do this for the user."
+            "The user is not aware of the different specialized assistants, so do not mention them; just quietly delegate through function calls. "
+            "Provide detailed information to the customer, and always double-check the database before concluding that information is unavailable. "
+            " When searching, be persistent. Expand your query bounds if the first search returns no results. "
+            " If a search comes up empty, expand your search before giving up."
+            "\n\nCurrent user flight information:\n<Flights>\n{user_info}\n</Flights>"
+            "\nCurrent time: {time}.",
+        ),
+        ("placeholder", "{messages}"),
+    ]
+).partial(time=datetime.now)
+primary_assistant_tools = [
+    # TavilySearchResults(tavily_api_key=settings.TAVILY_API_KEY, max_results=1),
+    search_flights,
+    lookup_policy,
+]
+assistant_runnable = primary_assistant_prompt | llm.bind_tools(
+    primary_assistant_tools
+    + [
+        ToFlightBookingAssistant,
+        ToBookCarRental,
+        ToHotelBookingAssistant,
+        ToBookExcursion,
+    ]
+)
