+++ cse-project/backend/app/services/cs_agents/requirements.txt
@@ -1,0 +1,46 @@
+# Core
+fastapi>=0.104.1,<0.105
+uvicorn>=0.24.0,<0.25
+python-dotenv>=1.0.0,<2.0.0
+pydantic>=2.7.0,<3.0.0  # Adjusted to satisfy all dependencies
+pydantic-settings>=2.0.3,<3.0.0  # Adjusted to satisfy all dependencies
+# Database
+motor>=3.3.1,<4.0.0
+redis>=5.0.1,<6.0.0
+pymilvus>=2.3.3,<3.0.0
+#sqlite3
+# AI/ML
+langchain>=0.0.330,<1.0.0
+langchain-community>=0.0.1,<1.0.0
+langchain-core>=0.0.13,<1.0.0
+langchain_openai>=0.2.4
+llama-index>=0.11.21,<1.0.0
+openai>=1.52.2
+langsmith>=0.1.140
+langgraph>=0.2.53
+langchain-anthropic>=0.2.4
+tavily-python>=0.5.0
+pandas>=2.2.3
+
+# WebSocket
+websockets>=12.0,<13.0
+
+# Testing
+pytest>=7.4.3,<8.0.0
+httpx>=0.25.1,<1.0.0
+
+# Development
+black>=23.10.1,<24.0.0
+isort>=5.12.0,<6.0.0
+flake8>=6.1.0,<7.0.0
+mypy>=1.6.1,<2.0.0
+pytest-cov>=4.1.0,<5.0.0
+pytest-asyncio>=0.21.1,<0.22.0
+
+pydantic>=2.9.1
+requests>=2.32.3  # Ensure requests is included
+
+# Additional Packages
+IPython>=8.0.0  # Ensure IPython is included
+typing_extensions>=4.0.0  # Ensure typing_extensions is included
+
