+++ cse-project/backend/app/services/cs_agents/routes.py
@@ -1,0 +1,91 @@
+from typing import Literal
+from langgraph.prebuilt import tools_condition
+from langgraph.graph import END
+from app.services.cs_agents.assistants.common import CompleteOrEscalate
+from app.services.cs_agents.assistants.primary_assistant import ToFlightBookingAssistant, ToBookCarRental, ToHotelBookingAssistant, ToBookExcursion
+from app.services.cs_agents.state import State
+from app.services.cs_agents.assistants.flight_booking_assistant import update_flight_safe_tools
+from app.services.cs_agents.assistants.hotel_booking_assistant import book_hotel_safe_tools
+from app.services.cs_agents.assistants.car_rental_assistant import book_car_rental_safe_tools
+from app.services.cs_agents.assistants.excursion_assistant import book_excursion_safe_tools
+def route_update_flight(state: State):
+    route = tools_condition(state)
+    if route == END:
+        return END
+    tool_calls = state["messages"][-1].tool_calls
+    did_cancel = any(tc["name"] == CompleteOrEscalate.__name__ for tc in tool_calls)
+    if did_cancel:
+        return "leave_skill"
+    safe_toolnames = [t.name for t in update_flight_safe_tools]
+    if all(tc["name"] in safe_toolnames for tc in tool_calls):
+        return "update_flight_safe_tools"
+    return "update_flight_sensitive_tools"
+
+def route_book_car_rental(state: State):
+    route = tools_condition(state)
+    if route == END:
+        return END
+    tool_calls = state["messages"][-1].tool_calls
+    did_cancel = any(tc["name"] == CompleteOrEscalate.__name__ for tc in tool_calls)
+    if did_cancel:
+        return "leave_skill"
+    safe_toolnames = [t.name for t in book_car_rental_safe_tools]
+    if all(tc["name"] in safe_toolnames for tc in tool_calls):
+        return "book_car_rental_safe_tools"
+    return "book_car_rental_sensitive_tools"
+
+def route_book_hotel(state: State):
+    route = tools_condition(state)
+    if route == END:
+        return END
+    tool_calls = state["messages"][-1].tool_calls
+    did_cancel = any(tc["name"] == CompleteOrEscalate.__name__ for tc in tool_calls)
+    if did_cancel:
+        return "leave_skill"
+    tool_names = [t.name for t in book_hotel_safe_tools]
+    if all(tc["name"] in tool_names for tc in tool_calls):
+        return "book_hotel_safe_tools"
+    return "book_hotel_sensitive_tools"
+
+def route_book_excursion(state: State):
+    route = tools_condition(state)
+    if route == END:
+        return END
+    tool_calls = state["messages"][-1].tool_calls
+    did_cancel = any(tc["name"] == CompleteOrEscalate.__name__ for tc in tool_calls)
+    if did_cancel:
+        return "leave_skill"
+    tool_names = [t.name for t in book_excursion_safe_tools]
+    if all(tc["name"] in tool_names for tc in tool_calls):
+        return "book_excursion_safe_tools"
+    return "book_excursion_sensitive_tools"
+
+def route_primary_assistant(state: State):
+    route = tools_condition(state)
+    if route == END:
+        return END
+    tool_calls = state["messages"][-1].tool_calls
+    if tool_calls:
+        if tool_calls[0]["name"] == ToFlightBookingAssistant.__name__:
+            return "enter_update_flight"
+        elif tool_calls[0]["name"] == ToBookCarRental.__name__:
+            return "enter_book_car_rental"
+        elif tool_calls[0]["name"] == ToHotelBookingAssistant.__name__:
+            return "enter_book_hotel"
+        elif tool_calls[0]["name"] == ToBookExcursion.__name__:
+            return "enter_book_excursion"
+        return "primary_assistant_tools"
+    raise ValueError("Invalid route")
+
+def route_to_workflow(state: State) -> Literal[
+    "primary_assistant",
+    "update_flight",
+    "book_car_rental",
+    "book_hotel",
+    "book_excursion",
+]:
+    """If we are in a delegated state, route directly to the appropriate assistant."""
+    dialog_state = state.get("dialog_state")
+    if not dialog_state:
+        return "primary_assistant"
+    return dialog_state[-1]
