+++ cse-project/backend/app/services/cs_agents/tools/car_rental_tools.py
@@ -1,0 +1,143 @@
+import sqlite3
+from datetime import date, datetime
+from typing import Optional, Union
+from langchain_core.tools import tool
+from app.services.cs_agents.utils import ensure_db_accessible
+
+db = "/Users/y001850/TENET/TENET/cse-project/services/travel2.sqlite"
+
+@tool
+def search_car_rentals(
+    location: Optional[str] = None,
+    name: Optional[str] = None,
+    price_tier: Optional[str] = None,
+    start_date: Optional[Union[datetime, date]] = None,
+    end_date: Optional[Union[datetime, date]] = None,
+) -> list[dict]:
+    """
+    Search for car rentals based on location, name, price tier, start date, and end date.
+
+    Args:
+        location (Optional[str]): The location of the car rental. Defaults to None.
+        name (Optional[str]): The name of the car rental company. Defaults to None.
+        price_tier (Optional[str]): The price tier of the car rental. Defaults to None.
+        start_date (Optional[Union[datetime, date]]): The start date of the car rental. Defaults to None.
+        end_date (Optional[Union[datetime, date]]): The end date of the car rental. Defaults to None.
+
+    Returns:
+        list[dict]: A list of car rental dictionaries matching the search criteria.
+    """
+    ensure_db_accessible(db)
+    conn = sqlite3.connect(db)
+    cursor = conn.cursor()
+
+    query = "SELECT * FROM car_rentals WHERE 1=1"
+    params = []
+
+    if location:
+        query += " AND location LIKE ?"
+        params.append(f"%{location}%")
+    if name:
+        query += " AND name LIKE ?"
+        params.append(f"%{name}%")
+    # For our tutorial, we will let you match on any dates and price tier.
+    # (since our toy dataset doesn't have much data)
+    cursor.execute(query, params)
+    results = cursor.fetchall()
+
+    conn.close()
+
+    return [
+        dict(zip([column[0] for column in cursor.description], row)) for row in results
+    ]
+
+@tool
+def book_car_rental(rental_id: int) -> str:
+    """
+    Book a car rental by its ID.
+
+    Args:
+        rental_id (int): The ID of the car rental to book.
+
+    Returns:
+        str: A message indicating whether the car rental was successfully booked or not.
+    """
+    ensure_db_accessible(db)
+    conn = sqlite3.connect(db)
+    cursor = conn.cursor()
+
+    cursor.execute("UPDATE car_rentals SET booked = 1 WHERE id = ?", (rental_id,))
+    conn.commit()
+
+    if cursor.rowcount > 0:
+        conn.close()
+        return f"Car rental {rental_id} successfully booked."
+    else:
+        conn.close()
+        return f"No car rental found with ID {rental_id}."
+
+@tool
+def update_car_rental(
+    rental_id: int,
+    start_date: Optional[Union[datetime, date]] = None,
+    end_date: Optional[Union[datetime, date]] = None,
+) -> str:
+    """
+    Update a car rental's start and end dates by its ID.
+
+    Args:
+        rental_id (int): The ID of the car rental to update.
+        start_date (Optional[Union[datetime, date]]): The new start date of the car rental. Defaults to None.
+        end_date (Optional[Union[datetime, date]]): The new end date of the car rental. Defaults to None.
+
+    Returns:
+        str: A message indicating whether the car rental was successfully updated or not.
+    """
+    ensure_db_accessible(db)
+    conn = sqlite3.connect(db)
+    cursor = conn.cursor()
+
+    if start_date:
+        cursor.execute(
+            "UPDATE car_rentals SET start_date = ? WHERE id = ?",
+            (start_date, rental_id),
+        )
+    if end_date:
+        cursor.execute(
+            "UPDATE car_rentals SET end_date = ? WHERE id = ?", (end_date, rental_id)
+        )
+
+    conn.commit()
+
+    if cursor.rowcount > 0:
+        conn.close()
+        return f"Car rental {rental_id} successfully updated."
+    else:
+        conn.close()
+        return f"No car rental found with ID {rental_id}."
+
+@tool
+def cancel_car_rental(rental_id: int) -> str:
+    
+    """
+    Cancel a car rental by its ID.
+
+    Args:
+        rental_id (int): The ID of the car rental to cancel.
+
+    Returns:
+        str: A message indicating whether the car rental was successfully cancelled or not.
+    """
+    ensure_db_accessible(db)
+    conn = sqlite3.connect(db)
+    cursor = conn.cursor()
+
+    cursor.execute("UPDATE car_rentals SET booked = 0 WHERE id = ?", (rental_id,))
+    conn.commit()
+
+    if cursor.rowcount > 0:
+        conn.close()
+        return f"Car rental {rental_id} successfully cancelled."
+    else:
+        conn.close()
+        return f"No car rental found with ID {rental_id}."
