+++ cse-project/backend/app/services/cs_agents/utils.py
@@ -1,0 +1,95 @@
+from typing import Callable
+from langchain_core.messages import ToolMessage
+from langchain_core.runnables import RunnableLambda
+from langgraph.prebuilt import ToolNode
+from app.services.cs_agents.state import State
+import sqlite3
+import numpy as np
+class VectorStoreRetriever:
+    def __init__(self, docs: list, vectors: list, oai_client):
+        self._arr = np.array(vectors)
+        self._docs = docs
+        self._client = oai_client
+
+    @classmethod
+    def from_docs(cls, docs, oai_client):
+        embeddings = oai_client.embeddings.create(
+            model="text-embedding-3-small", input=[doc["page_content"] for doc in docs]
+        )
+        vectors = [emb.embedding for emb in embeddings.data]
+        return cls(docs, vectors, oai_client)
+
+    def query(self, query: str, k: int = 5) -> list[dict]:
+        embed = self._client.embeddings.create(
+            model="text-embedding-3-small", input=[query]
+        )
+        # "@" is just a matrix multiplication in python
+        scores = np.array(embed.data[0].embedding) @ self._arr.T
+        top_k_idx = np.argpartition(scores, -k)[-k:]
+        top_k_idx_sorted = top_k_idx[np.argsort(-scores[top_k_idx])]
+        return [
+            {**self._docs[idx], "similarity": scores[idx]} for idx in top_k_idx_sorted
+        ]
+
+def create_entry_node(assistant_name: str, new_dialog_state: str) -> Callable:
+    def entry_node(state: State) -> dict:
+        tool_call_id = state["messages"][-1].tool_calls[0]["id"]
+        return {
+            "messages": [
+                ToolMessage(
+                    content=f"The assistant is now the {assistant_name}. Reflect on the above conversation between the host assistant and the user."
+                    f" The user's intent is unsatisfied. Use the provided tools to assist the user. Remember, you are {assistant_name},"
+                    " and the booking, update, other other action is not complete until after you have successfully invoked the appropriate tool."
+                    " If the user changes their mind or needs help for other tasks, call the CompleteOrEscalate function to let the primary host assistant take control."
+                    " Do not mention who you are - just act as the proxy for the assistant.",
+                    tool_call_id=tool_call_id,
+                )
+            ],
+            "dialog_state": new_dialog_state,
+        }
+
+    return entry_node
+
+def handle_tool_error(state) -> dict:
+    error = state.get("error")
+    tool_calls = state["messages"][-1].tool_calls
+    return {
+        "messages": [
+            ToolMessage(
+                content=f"Error: {repr(error)}\n please fix your mistakes.",
+                tool_call_id=tc["id"],
+            )
+            for tc in tool_calls
+        ]
+    }
+
+def create_tool_node_with_fallback(tools: list) -> dict:
+    return ToolNode(tools).with_fallbacks(
+        [RunnableLambda(handle_tool_error)], exception_key="error"
+    )
+
+def _print_event(event: dict, _printed: set, max_length=1500):
+    current_state = event.get("dialog_state")
+    if current_state:
+        print("Currently in: ", current_state[-1])
+    message = event.get("messages")
+    if message:
+        if isinstance(message, list):
+            message = message[-1]
+        if message.id not in _printed:
+            msg_repr = message.pretty_repr(html=True)
+            if len(msg_repr) > max_length:
+                msg_repr = msg_repr[:max_length] + " ... (truncated)"
+            print(msg_repr)
+            _printed.add(message.id)
+
+def ensure_db_accessible(db_path: str):
+    """Ensure the database is accessible."""
+    try:
+        conn = sqlite3.connect(db_path)
+        cursor = conn.cursor()
+        cursor.execute("SELECT 1")
+        cursor.close()
+        conn.close()
+    except sqlite3.OperationalError as e:
+        raise RuntimeError("Database is not accessible") from e
