--- cse-project/backend/app/services/graph.py
@@ -1,55 +1,0 @@
-from langgraph.checkpoint.memory import MemorySaver
-from langgraph.graph import END, StateGraph, START
-from langgraph.prebuilt import tools_condition
-from app.services.agent import Assistant, part_3_assistant_runnable, fetch_user_flight_information, part_3_safe_tools, part_3_sensitive_tools, sensitive_tool_names
-from app.services.utilities import create_tool_node_with_fallback
-from app.services.state import State
-from IPython.display import Image, display
-
-builder = StateGraph(State)
-
-def user_info(state: State):
-    return {"user_info": fetch_user_flight_information.invoke({})}
-
-
-# NEW: The fetch_user_info node runs first, meaning our assistant can see the user's flight information without
-# having to take an action
-builder.add_node("fetch_user_info", user_info)
-builder.add_edge(START, "fetch_user_info")
-builder.add_node("assistant", Assistant(part_3_assistant_runnable))
-builder.add_node("safe_tools", create_tool_node_with_fallback(part_3_safe_tools))
-builder.add_node(
-    "sensitive_tools", create_tool_node_with_fallback(part_3_sensitive_tools)
-)
-# Define logic
-builder.add_edge("fetch_user_info", "assistant")
-
-
-def route_tools(state: State):
-    next_node = tools_condition(state)
-    # If no tools are invoked, return to the user
-    if next_node == END:
-        return END
-    ai_message = state["messages"][-1]
-    # This assumes single tool calls. To handle parallel tool calling, you'd want to
-    # use an ANY condition
-    first_tool_call = ai_message.tool_calls[0]
-    if first_tool_call["name"] in sensitive_tool_names:
-        return "sensitive_tools"
-    return "safe_tools"
-
-
-builder.add_conditional_edges(
-    "assistant", route_tools, ["safe_tools", "sensitive_tools", END]
-)
-builder.add_edge("safe_tools", "assistant")
-builder.add_edge("sensitive_tools", "assistant")
-
-memory = MemorySaver()
-part_3_graph = builder.compile(
-    checkpointer=memory,
-    # NEW: The graph will always halt before executing the "tools" node.
-    # The user can approve or reject (or even alter the request) before
-    # the assistant continues
-    interrupt_before=["sensitive_tools"],
-)
