+++ cse-project/backend/app/services/langgraph.json
@@ -1,0 +1,9 @@
+{
+  "graphs": {
+    "agent": "./cs_agents/agent.py:part_4_graph"
+  },
+  "env": "../../.env",
+  "dependencies": [
+    "./cs_agents"
+  ]
+}
