--- cse-project/backend/app/services/state.py
@@ -1,10 +1,0 @@
-from typing import Annotated
-
-from typing_extensions import TypedDict
-
-from langgraph.graph.message import AnyMessage, add_messages
-
-
-class State(TypedDict):
-    messages: Annotated[list[AnyMessage], add_messages]
-    user_info: str
