--- cse-project/backend/app/services/tools.py
@@ -1,650 +1,0 @@
-import os
-from datetime import date, datetime
-from typing import Optional, Union
-import re
-import numpy as np
-import openai
-from langchain_core.tools import tool
-import sqlite3
-import pytz
-from langchain_core.runnables import RunnableConfig
-import requests
-from app.core.config import settings
-
-db = "/Users/y001850/TENET/TENET/cse-project/services/travel2.sqlite"
-response = requests.get(
-    "https://storage.googleapis.com/benchmarks-artifacts/travel-db/swiss_faq.md"
-)
-response.raise_for_status()
-faq_text = response.text
-
-docs = [{"page_content": txt} for txt in re.split(r"(?=\n##)", faq_text)]
-
-def ensure_db_accessible(db_path):
-    if not os.path.exists(db_path):
-        raise FileNotFoundError(f"Database file {db_path} does not exist.")
-    if not os.access(db_path, os.R_OK | os.W_OK):
-        raise PermissionError(f"Database file {db_path} is not accessible.")
-
-class VectorStoreRetriever:
-    def __init__(self, docs: list, vectors: list, oai_client):
-        self._arr = np.array(vectors)
-        self._docs = docs
-        self._client = oai_client
-
-    @classmethod
-    def from_docs(cls, docs, oai_client):
-        embeddings = oai_client.embeddings.create(
-            model="text-embedding-3-small", input=[doc["page_content"] for doc in docs]
-        )
-        vectors = [emb.embedding for emb in embeddings.data]
-        return cls(docs, vectors, oai_client)
-
-    def query(self, query: str, k: int = 5) -> list[dict]:
-        embed = self._client.embeddings.create(
-            model="text-embedding-3-small", input=[query]
-        )
-        # "@" is just a matrix multiplication in python
-        scores = np.array(embed.data[0].embedding) @ self._arr.T
-        top_k_idx = np.argpartition(scores, -k)[-k:]
-        top_k_idx_sorted = top_k_idx[np.argsort(-scores[top_k_idx])]
-        return [
-            {**self._docs[idx], "similarity": scores[idx]} for idx in top_k_idx_sorted
-        ]
-
-
-retriever = VectorStoreRetriever.from_docs(docs, openai.Client(api_key=settings.OPENAI_API_KEY))
-
-
-@tool
-def lookup_policy(query: str) -> str:
-    """Consult the company policies to check whether certain options are permitted.
-    Use this before making any flight changes performing other 'write' events."""
-    docs = retriever.query(query, k=2)
-    return "\n\n".join([doc["page_content"] for doc in docs])
-
-@tool
-def search_car_rentals(
-    location: Optional[str] = None,
-    name: Optional[str] = None,
-    price_tier: Optional[str] = None,
-    start_date: Optional[Union[datetime, date]] = None,
-    end_date: Optional[Union[datetime, date]] = None,
-) -> list[dict]:
-    """
-    Search for car rentals based on location, name, price tier, start date, and end date.
-
-    Args:
-        location (Optional[str]): The location of the car rental. Defaults to None.
-        name (Optional[str]): The name of the car rental company. Defaults to None.
-        price_tier (Optional[str]): The price tier of the car rental. Defaults to None.
-        start_date (Optional[Union[datetime, date]]): The start date of the car rental. Defaults to None.
-        end_date (Optional[Union[datetime, date]]): The end date of the car rental. Defaults to None.
-
-    Returns:
-        list[dict]: A list of car rental dictionaries matching the search criteria.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    query = "SELECT * FROM car_rentals WHERE 1=1"
-    params = []
-
-    if location:
-        query += " AND location LIKE ?"
-        params.append(f"%{location}%")
-    if name:
-        query += " AND name LIKE ?"
-        params.append(f"%{name}%")
-    # For our tutorial, we will let you match on any dates and price tier.
-    # (since our toy dataset doesn't have much data)
-    cursor.execute(query, params)
-    results = cursor.fetchall()
-
-    conn.close()
-
-    return [
-        dict(zip([column[0] for column in cursor.description], row)) for row in results
-    ]
-
-@tool
-def book_car_rental(rental_id: int) -> str:
-    """
-    Book a car rental by its ID.
-
-    Args:
-        rental_id (int): The ID of the car rental to book.
-
-    Returns:
-        str: A message indicating whether the car rental was successfully booked or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute("UPDATE car_rentals SET booked = 1 WHERE id = ?", (rental_id,))
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Car rental {rental_id} successfully booked."
-    else:
-        conn.close()
-        return f"No car rental found with ID {rental_id}."
-
-@tool
-def update_car_rental(
-    rental_id: int,
-    start_date: Optional[Union[datetime, date]] = None,
-    end_date: Optional[Union[datetime, date]] = None,
-) -> str:
-    """
-    Update a car rental's start and end dates by its ID.
-
-    Args:
-        rental_id (int): The ID of the car rental to update.
-        start_date (Optional[Union[datetime, date]]): The new start date of the car rental. Defaults to None.
-        end_date (Optional[Union[datetime, date]]): The new end date of the car rental. Defaults to None.
-
-    Returns:
-        str: A message indicating whether the car rental was successfully updated or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    if start_date:
-        cursor.execute(
-            "UPDATE car_rentals SET start_date = ? WHERE id = ?",
-            (start_date, rental_id),
-        )
-    if end_date:
-        cursor.execute(
-            "UPDATE car_rentals SET end_date = ? WHERE id = ?", (end_date, rental_id)
-        )
-
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Car rental {rental_id} successfully updated."
-    else:
-        conn.close()
-        return f"No car rental found with ID {rental_id}."
-
-@tool
-def cancel_car_rental(rental_id: int) -> str:
-    
-    """
-    Cancel a car rental by its ID.
-
-    Args:
-        rental_id (int): The ID of the car rental to cancel.
-
-    Returns:
-        str: A message indicating whether the car rental was successfully cancelled or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute("UPDATE car_rentals SET booked = 0 WHERE id = ?", (rental_id,))
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Car rental {rental_id} successfully cancelled."
-    else:
-        conn.close()
-        return f"No car rental found with ID {rental_id}."
-
-@tool
-def search_trip_recommendations(
-    location: Optional[str] = None,
-    name: Optional[str] = None,
-    keywords: Optional[str] = None,
-) -> list[dict]:
-    """
-    Search for trip recommendations based on location, name, and keywords.
-
-    Args:
-        location (Optional[str]): The location of the trip recommendation. Defaults to None.
-        name (Optional[str]): The name of the trip recommendation. Defaults to None.
-        keywords (Optional[str]): The keywords associated with the trip recommendation. Defaults to None.
-
-    Returns:
-        list[dict]: A list of trip recommendation dictionaries matching the search criteria.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    query = "SELECT * FROM trip_recommendations WHERE 1=1"
-    params = []
-
-    if location:
-        query += " AND location LIKE ?"
-        params.append(f"%{location}%")
-    if name:
-        query += " AND name LIKE ?"
-        params.append(f"%{name}%")
-    if keywords:
-        keyword_list = keywords.split(",")
-        keyword_conditions = " OR ".join(["keywords LIKE ?" for _ in keyword_list])
-        query += f" AND ({keyword_conditions})"
-        params.extend([f"%{keyword.strip()}%" for keyword in keyword_list])
-
-    cursor.execute(query, params)
-    results = cursor.fetchall()
-
-    conn.close()
-
-    return [
-        dict(zip([column[0] for column in cursor.description], row)) for row in results
-    ]
-
-@tool
-def book_excursion(recommendation_id: int) -> str:
-    """
-    Book a excursion by its recommendation ID.
-
-    Args:
-        recommendation_id (int): The ID of the trip recommendation to book.
-
-    Returns:
-        str: A message indicating whether the trip recommendation was successfully booked or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute(
-        "UPDATE trip_recommendations SET booked = 1 WHERE id = ?", (recommendation_id,)
-    )
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Trip recommendation {recommendation_id} successfully booked."
-    else:
-        conn.close()
-        return f"No trip recommendation found with ID {recommendation_id}."
-
-@tool
-def update_excursion(recommendation_id: int, details: str) -> str:
-    """
-    Update a trip recommendation's details by its ID.
-
-    Args:
-        recommendation_id (int): The ID of the trip recommendation to update.
-        details (str): The new details of the trip recommendation.
-
-    Returns:
-        str: A message indicating whether the trip recommendation was successfully updated or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute(
-        "UPDATE trip_recommendations SET details = ? WHERE id = ?",
-        (details, recommendation_id),
-    )
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Trip recommendation {recommendation_id} successfully updated."
-    else:
-        conn.close()
-        return f"No trip recommendation found with ID {recommendation_id}."
-
-@tool
-def cancel_excursion(recommendation_id: int) -> str:
-    """
-    Cancel a trip recommendation by its ID.
-
-    Args:
-        recommendation_id (int): The ID of the trip recommendation to cancel.
-
-    Returns:
-        str: A message indicating whether the trip recommendation was successfully cancelled or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute(
-        "UPDATE trip_recommendations SET booked = 0 WHERE id = ?", (recommendation_id,)
-    )
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Trip recommendation {recommendation_id} successfully cancelled."
-    else:
-        conn.close()
-        return f"No trip recommendation found with ID {recommendation_id}."
-
-@tool
-def fetch_user_flight_information(config: RunnableConfig) -> list[dict]:
-    """Fetch all tickets for the user along with corresponding flight information and seat assignments.
-
-    Returns:
-        A list of dictionaries where each dictionary contains the ticket details,
-        associated flight details, and the seat assignments for each ticket belonging to the user.
-    """
-    configuration = config.get("configurable", {})
-    passenger_id = configuration.get("passenger_id", None)
-    if not passenger_id:
-        raise ValueError("No passenger ID configured.")
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    query = """
-    SELECT 
-        t.ticket_no, t.book_ref,
-        f.flight_id, f.flight_no, f.departure_airport, f.arrival_airport, f.scheduled_departure, f.scheduled_arrival,
-        bp.seat_no, tf.fare_conditions
-    FROM 
-        tickets t
-        JOIN ticket_flights tf ON t.ticket_no = tf.ticket_no
-        JOIN flights f ON tf.flight_id = f.flight_id
-        JOIN boarding_passes bp ON bp.ticket_no = t.ticket_no AND bp.flight_id = f.flight_id
-    WHERE 
-        t.passenger_id = ?
-    """
-    cursor.execute(query, (passenger_id,))
-    rows = cursor.fetchall()
-    column_names = [column[0] for column in cursor.description]
-    results = [dict(zip(column_names, row)) for row in rows]
-
-    cursor.close()
-    conn.close()
-
-    return results
-
-@tool
-def search_flights(
-    departure_airport: Optional[str] = None,
-    arrival_airport: Optional[str] = None,
-    start_time: Optional[date | datetime] = None,
-    end_time: Optional[date | datetime] = None,
-    limit: int = 20,
-) -> list[dict]:
-    """Search for flights based on departure airport, arrival airport, and departure time range."""
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    query = "SELECT * FROM flights WHERE 1 = 1"
-    params = []
-
-    if departure_airport:
-        query += " AND departure_airport = ?"
-        params.append(departure_airport)
-
-    if arrival_airport:
-        query += " AND arrival_airport = ?"
-        params.append(arrival_airport)
-
-    if start_time:
-        query += " AND scheduled_departure >= ?"
-        params.append(start_time)
-
-    if end_time:
-        query += " AND scheduled_departure <= ?"
-        params.append(end_time)
-    query += " LIMIT ?"
-    params.append(limit)
-    cursor.execute(query, params)
-    rows = cursor.fetchall()
-    column_names = [column[0] for column in cursor.description]
-    results = [dict(zip(column_names, row)) for row in rows]
-
-    cursor.close()
-    conn.close()
-
-    return results
-
-@tool
-def update_ticket_to_new_flight(
-    ticket_no: str, new_flight_id: int, *, config: RunnableConfig
-) -> str:
-    """Update the user's ticket to a new valid flight."""
-    configuration = config.get("configurable", {})
-    passenger_id = configuration.get("passenger_id", None)
-    if not passenger_id:
-        raise ValueError("No passenger ID configured.")
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute(
-        "SELECT departure_airport, arrival_airport, scheduled_departure FROM flights WHERE flight_id = ?",
-        (new_flight_id,),
-    )
-    new_flight = cursor.fetchone()
-    if not new_flight:
-        cursor.close()
-        conn.close()
-        return "Invalid new flight ID provided."
-    column_names = [column[0] for column in cursor.description]
-    new_flight_dict = dict(zip(column_names, new_flight))
-    timezone = pytz.timezone("Etc/GMT-3")
-    current_time = datetime.now(tz=timezone)
-    departure_time = datetime.strptime(
-        new_flight_dict["scheduled_departure"], "%Y-%m-%d %H:%M:%S.%f%z"
-    )
-    time_until = (departure_time - current_time).total_seconds()
-    if time_until < (3 * 3600):
-        return f"Not permitted to reschedule to a flight that is less than 3 hours from the current time. Selected flight is at {departure_time}."
-
-    cursor.execute(
-        "SELECT flight_id FROM ticket_flights WHERE ticket_no = ?", (ticket_no,)
-    )
-    current_flight = cursor.fetchone()
-    if not current_flight:
-        cursor.close()
-        conn.close()
-        return "No existing ticket found for the given ticket number."
-
-    # Check the signed-in user actually has this ticket
-    cursor.execute(
-        "SELECT * FROM tickets WHERE ticket_no = ? AND passenger_id = ?",
-        (ticket_no, passenger_id),
-    )
-    current_ticket = cursor.fetchone()
-    if not current_ticket:
-        cursor.close()
-        conn.close()
-        return f"Current signed-in passenger with ID {passenger_id} not the owner of ticket {ticket_no}"
-
-    # In a real application, you'd likely add additional checks here to enforce business logic,
-    # like "does the new departure airport match the current ticket", etc.
-    # While it's best to try to be *proactive* in 'type-hinting' policies to the LLM
-    # it's inevitably going to get things wrong, so you **also** need to ensure your
-    # API enforces valid behavior
-    cursor.execute(
-        "UPDATE ticket_flights SET flight_id = ? WHERE ticket_no = ?",
-        (new_flight_id, ticket_no),
-    )
-    conn.commit()
-
-    cursor.close()
-    conn.close()
-    return "Ticket successfully updated to new flight."
-
-@tool
-def cancel_ticket(ticket_no: str, *, config: RunnableConfig) -> str:
-    """Cancel the user's ticket and remove it from the database."""
-    configuration = config.get("configurable", {})
-    passenger_id = configuration.get("passenger_id", None)
-    if not passenger_id:
-        raise ValueError("No passenger ID configured.")
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute(
-        "SELECT flight_id FROM ticket_flights WHERE ticket_no = ?", (ticket_no,)
-    )
-    existing_ticket = cursor.fetchone()
-    if not existing_ticket:
-        cursor.close()
-        conn.close()
-        return "No existing ticket found for the given ticket number."
-
-    # Check the signed-in user actually has this ticket
-    cursor.execute(
-        "SELECT flight_id FROM tickets WHERE ticket_no = ? AND passenger_id = ?",
-        (ticket_no, passenger_id),
-    )
-    current_ticket = cursor.fetchone()
-    if not current_ticket:
-        cursor.close()
-        conn.close()
-        return f"Current signed-in passenger with ID {passenger_id} not the owner of ticket {ticket_no}"
-
-    cursor.execute("DELETE FROM ticket_flights WHERE ticket_no = ?", (ticket_no,))
-    conn.commit()
-
-    cursor.close()
-    conn.close()
-    return "Ticket successfully cancelled."
-
-@tool
-def search_hotels(
-    location: Optional[str] = None,
-    name: Optional[str] = None,
-    price_tier: Optional[str] = None,
-    checkin_date: Optional[Union[datetime, date]] = None,
-    checkout_date: Optional[Union[datetime, date]] = None,
-) -> list[dict]:
-    """
-    Search for hotels based on location, name, price tier, check-in date, and check-out date.
-
-    Args:
-        location (Optional[str]): The location of the hotel. Defaults to None.
-        name (Optional[str]): The name of the hotel. Defaults to None.
-        price_tier (Optional[str]): The price tier of the hotel. Defaults to None. Examples: Midscale, Upper Midscale, Upscale, Luxury
-        checkin_date (Optional[Union[datetime, date]]): The check-in date of the hotel. Defaults to None.
-        checkout_date (Optional[Union[datetime, date]]): The check-out date of the hotel. Defaults to None.
-
-    Returns:
-        list[dict]: A list of hotel dictionaries matching the search criteria.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    query = "SELECT * FROM hotels WHERE 1=1"
-    params = []
-
-    if location:
-        query += " AND location LIKE ?"
-        params.append(f"%{location}%")
-    if name:
-        query += " AND name LIKE ?"
-        params.append(f"%{name}%")
-    # For the sake of this tutorial, we will let you match on any dates and price tier.
-    cursor.execute(query, params)
-    results = cursor.fetchall()
-
-    conn.close()
-
-    return [
-        dict(zip([column[0] for column in cursor.description], row)) for row in results
-    ]
-
-@tool
-def book_hotel(hotel_id: int) -> str:
-    """
-    Book a hotel by its ID.
-
-    Args:
-        hotel_id (int): The ID of the hotel to book.
-
-    Returns:
-        str: A message indicating whether the hotel was successfully booked or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute("UPDATE hotels SET booked = 1 WHERE id = ?", (hotel_id,))
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Hotel {hotel_id} successfully booked."
-    else:
-        conn.close()
-        return f"No hotel found with ID {hotel_id}."
-
-@tool
-def update_hotel(
-    hotel_id: int,
-    checkin_date: Optional[Union[datetime, date]] = None,
-    checkout_date: Optional[Union[datetime, date]] = None,
-) -> str:
-    """
-    Update a hotel's check-in and check-out dates by its ID.
-
-    Args:
-        hotel_id (int): The ID of the hotel to update.
-        checkin_date (Optional[Union[datetime, date]]): The new check-in date of the hotel. Defaults to None.
-        checkout_date (Optional[Union[datetime, date]]): The new check-out date of the hotel. Defaults to None.
-
-    Returns:
-        str: A message indicating whether the hotel was successfully updated or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    if checkin_date:
-        cursor.execute(
-            "UPDATE hotels SET checkin_date = ? WHERE id = ?", (checkin_date, hotel_id)
-        )
-    if checkout_date:
-        cursor.execute(
-            "UPDATE hotels SET checkout_date = ? WHERE id = ?",
-            (checkout_date, hotel_id),
-        )
-
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Hotel {hotel_id} successfully updated."
-    else:
-        conn.close()
-        return f"No hotel found with ID {hotel_id}."
-
-@tool
-def cancel_hotel(hotel_id: int) -> str:
-    """
-    Cancel a hotel by its ID.
-
-    Args:
-        hotel_id (int): The ID of the hotel to cancel.
-
-    Returns:
-        str: A message indicating whether the hotel was successfully cancelled or not.
-    """
-    ensure_db_accessible(db)
-    conn = sqlite3.connect(db)
-    cursor = conn.cursor()
-
-    cursor.execute("UPDATE hotels SET booked = 0 WHERE id = ?", (hotel_id,))
-    conn.commit()
-
-    if cursor.rowcount > 0:
-        conn.close()
-        return f"Hotel {hotel_id} successfully cancelled."
-    else:
-        conn.close()
-        return f"No hotel found with ID {hotel_id}."
